Inline introduction.adoc
Having it in a separate file confuses VSCode's Asciidoc support a bit because the heading sequence isn't correct - it goes H1->H3. I think it's also less confusing for humans this way, because only the first part of the introduction was in the introduction.adoc, and it hardly seems worth having a separate file for one paragraph.